adicao da funcao delete

+++ src/delete.php
@@ -1,12 +1,13 @@
 <?php
 require 'db.php';
 
-// obter o id da tarefa
-$id = $_GET['id'] ?? null;
-
-if ($id) {
-  $stmt = $pdo->prepare("DELETE FROM tarefas WHERE id = ?");
-  $stmt->execute([$ID]);
+// verificacao da requisicao feita com post no index
+if ($_SERVER["REQUEST_METHOD"] === "POST") {
+    $id = $_POST['id'] ?? null;
+    if ($id) {
+        $stmt = $pdo->prepare("DELETE FROM tarefas WHERE id = ?");
+        $stmt->execute([$id]);
+    }
 }
-
-header('Location: index.php');
+header("Location: index.php");
+exit;
